src/requirements.txt
# Define the required python packages
pillow>=5.2.0
requests>=2.20.0
flask>=1.0.2
Flask-RESTful>=0.3.7
torchvision>=0.3.0